requirements.txt
# Project dependencies

flask==3.0.0
python-dotenv==1.0.0
supabase==2.3.0
requests==2.31.0
skyfield==1.48
bcrypt==4.1.2
werkzeug==3.0.1
